% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedRF.R
\name{MixedRF}
\alias{MixedRF}
\title{Mixed Random Forest}
\usage{
MixedRF(data, weights, alphas, k = 5, ntree = 100, prob = 0.8)
}
\arguments{
\item{data}{A data frame that contains the data, with the response variable y being a binary variable of 0 and 1 with 1 being the minority class.}

\item{weights}{A vector that contains the weights that we want to test on. The function uses k-fold CV to find best weight.}

\item{alphas}{A vector that contains the alphas that we want to test on. The function uses k-fold CV to find the best alpha.}

\item{k}{Indicates the number of folds to create during cross validation process. The default is set to 5.}

\item{ntree}{The number of trees used in the balanced random forest model. The default is set to 100.}

\item{prob}{The ratio of training set to the whole data. The default is set to 0.8}
}
\value{
A vector of length 2 that contains the fitted mixed random forest model and the confusion matrix using the test data.
}
\description{
This function utilizes the randomForest library to implement a mixed random forest.
The mixed random forest takes a parameter alpha and does downsampling on the original data so that the ratio is 1:alpha.
Then it uses weighted random forest for the downsampled data to fit a model. Finally, it returns the fitted model and the confusion matrix of the test data.
}
\examples{
# Example Usage
data <- caret::twoClassSim(1000, intercept = -15, linearVars = 15, noiseVars = 5)
names(data)[names(data)=="Class"]="y"
data$y = as.factor(ifelse(data$y=="Class1",0,1))

result <- MixedRF(data,weights=c(100,1000,10000,100000),alphas=c(2,3,4,5),k=5,ntree=100,prob=0.8)


}
