% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdRF.R
\name{ThresholdRF}
\alias{ThresholdRF}
\title{Threshold Random Forest}
\usage{
ThresholdRF(data, ntree = 3000, prob = 0.8)
}
\arguments{
\item{data}{A data frame that contains the data, with the response variable y being a binary variable of 0 and 1 with 1 being the minority class.}

\item{ntree}{The number of trees used in the balanced random forest model. The default is set to 3000.}

\item{prob}{The ratio of training set to the whole data. The default is set to 0.8}
}
\value{
A vector of length 2 that contains the fitted mixed random forest model and the confusion matrix using the test data.
}
\description{
Threshold Random Forest
}
\examples{

# Example Usage
data <- caret::twoClassSim(1000, intercept = -15, linearVars = 15, noiseVars = 5)
names(data)[names(data)=="Class"]="y"
data$y = as.factor(ifelse(data$y=="Class1",0,1))

result <- ThresholdRF(data,ntree=100,prob=0.8)

}
